Move everything h1 to h1
