+++ src/h1/client.rs
@@ -1,0 +1,140 @@
+use std::rc::Rc;
+
+use eyre::Context;
+use tracing::debug;
+
+use crate::{
+    bufpool::{AggBuf, IoChunkList},
+    io::ReadWriteOwned,
+    types::Request,
+    util::{read_and_parse, write_all_list},
+    Body, Response,
+};
+
+use super::{
+    body::{write_h1_body, BodyWriteMode, H1Body, H1BodyKind},
+    encode::encode_request,
+};
+
+pub struct ClientConf {}
+
+pub trait ClientDriver {
+    type Return;
+
+    async fn on_informational_response(&self, res: Response) -> eyre::Result<()>;
+    async fn on_final_response<B>(self, res: Response, body: B) -> eyre::Result<(B, Self::Return)>
+    where
+        B: Body;
+}
+
+/// Perform an HTTP/1.1 request against an HTTP/1.1 server
+///
+/// The transport will be returned unless the server requested connection close.
+pub async fn request<T, B, D>(
+    transport: Rc<T>,
+    req: Request,
+    body: B,
+    driver: D,
+) -> eyre::Result<(Option<Rc<T>>, B, D::Return)>
+where
+    T: ReadWriteOwned,
+    B: Body,
+    D: ClientDriver,
+{
+    // TODO: set content-length if body isn't empty / missing
+    let mode = if body.content_len().is_none() {
+        BodyWriteMode::Chunked
+    } else {
+        BodyWriteMode::ContentLength
+    };
+
+    let mut list = IoChunkList::default();
+    encode_request(req, &mut list)?;
+    _ = write_all_list(transport.as_ref(), list)
+        .await
+        .wrap_err("writing request headers")?;
+
+    // TODO: handle `expect: 100-continue` (don't start sending body until we get a 100 response)
+
+    let send_body_fut = {
+        let transport = transport.clone();
+        async move {
+            match write_h1_body(transport, body, mode).await {
+                Err(err) => {
+                    // TODO: find way to report this error to the driver without
+                    // spawning, without ref-counting the driver, etc.
+                    panic!("error writing request body: {err:?}");
+                }
+                Ok(body) => {
+                    debug!("done writing request body");
+                    Ok::<_, eyre::Report>(body)
+                }
+            }
+        }
+    };
+
+    let recv_res_fut = {
+        let transport = transport.clone();
+        async move {
+            let ups_buf = AggBuf::default();
+            let (ups_buf, res) = match read_and_parse(
+                super::parse::response,
+                transport.as_ref(),
+                ups_buf,
+                // TODO: make this configurable
+                64 * 1024,
+            )
+            .await
+            {
+                Ok(t) => match t {
+                    Some(t) => t,
+                    None => {
+                        return Err(eyre::eyre!(
+                            "server went away before sending response headers"
+                        ));
+                    }
+                },
+                Err(e) => {
+                    // TODO: return error instead?
+                    return Err(eyre::eyre!(
+                        "error reading response headers from server: {e:?}"
+                    ));
+                }
+            };
+            res.debug_print();
+
+            // TODO: handle informational responses
+
+            let chunked = res.headers.is_chunked_transfer_encoding();
+            // TODO: handle 204/304 separately
+            let content_len = res.headers.content_length().unwrap_or_default();
+
+            let res_body = H1Body {
+                transport: transport.clone(),
+                buf: ups_buf,
+                kind: if chunked {
+                    H1BodyKind::Chunked
+                } else if content_len > 0 {
+                    H1BodyKind::ContentLength(content_len)
+                } else {
+                    H1BodyKind::Empty
+                },
+                read: 0,
+                eof: false,
+            };
+
+            let conn_close = res.headers.is_connection_close();
+
+            let (res_body, ret) = driver.on_final_response(res, res_body).await?;
+            // TODO: re-use buffer for pooled connections?
+            drop(res_body);
+
+            Ok((ret, conn_close))
+        }
+    };
+
+    let (req_body, (ret, conn_close)) = tokio::try_join!(send_body_fut, recv_res_fut)?;
+
+    let transport = if conn_close { None } else { Some(transport) };
+    Ok((transport, req_body, ret))
+}
