+++ src/lib.rs
@@ -4,11 +4,18 @@
 #![feature(async_fn_in_trait)]
 #![feature(return_position_impl_trait_in_trait)]
 
-pub mod bufpool;
-pub mod io;
-pub mod parse;
-pub mod proto;
-pub mod types;
+mod util;
+
+mod bufpool;
+pub use bufpool::*;
+
+mod io;
+pub use io::*;
+
+mod types;
+pub use types::*;
+
+pub mod h1;
 
 /// re-exported so consumers can use whatever forked version we use
 pub use tokio_uring;
