--- src/proto/mod.rs
@@ -1,3 +1,0 @@
-pub mod errors;
-pub mod h1;
-pub mod util;
