+++ src/util.rs
@@ -5,7 +5,6 @@ use tracing::debug;
 use crate::{
     bufpool::{AggBuf, AggSlice, IoChunkList},
     io::{ReadOwned, WriteOwned},
-    proto::errors::SemanticError,
 };
 
 /// Returns `None` on EOF, error if partially parsed message.
@@ -88,4 +87,18 @@ pub(crate) async fn write_all_list(
 
     list.clear();
     Ok(list.into())
+}
+
+#[derive(thiserror::Error, Debug)]
+pub(crate) enum SemanticError {
+    #[error("the headers are too long")]
+    HeadersTooLong,
+}
+
+impl SemanticError {
+    pub(crate) fn as_http_response(&self) -> &'static [u8] {
+        match self {
+            Self::HeadersTooLong => b"HTTP/1.1 431 Request Header Fields Too Large\r\n\r\n",
+        }
+    }
 }
